Tweaked API a lot. Customizability should be more straightforward and transparent now.
Added a module-level compose() function that can take zero or more arguments.
The identity function is used if no arguments are passed.

Modified the README to omit some pedantic details that were mostly there as
filler in lieu of more useful content. Also modified the sample session to
include use as a decorator, and doctest of it from the commandline.