Reorganized directory structure to be more distribution-friendly. The tests are not distrubuted, so the bit in the README that invokes them is replaced by an interactive session that is short but is about the same. The docstring is the test, and that's available.
Added a link to toolz and removed the "see also" items pointing to as-
yet-unpublished packages that utilize this library.

Distribution builds, looks okay, but the upload fails. twine seems to
be bigly bugged. It won't run in a clean venv with no arguments. I'm
keen on doing a commit before taking a break from this. Honestly, the
motivation to get it all the way deployed is less now, might be a while
before I get back.